moved test from debugger test suite to remoteconfig suite
