+++ tests/internal/remoteconfig/test_remoteconfig.py
@@ -398,3 +398,33 @@ def test_remote_configuration_check_remote_config_enable_in_agent_errors(
     assert worker._state == worker._online if expected else worker._agent_check
     worker.stop_subscribers(True)
     worker.disable()
+
+
+@pytest.mark.subprocess(
+    parametrize=dict(
+        DD_REMOTE_CONFIGURATION_ENABLED=["1", "0"],
+    ),
+)
+def test_rc_default_products_registered():
+    """
+    By default, RC should be enabled. When RC is enabled, we will always
+    enable the tracer flare feature as well. There should be three products
+    registered when DD_REMOTE_CONFIGURATION_ENABLED is True
+    """
+    import os
+
+    from ddtrace.internal.utils.formats import asbool
+
+    rc_enabled = asbool(os.environ.get("DD_REMOTE_CONFIGURATION_ENABLED"))
+
+    # Import this to trigger the preload
+    from ddtrace import config
+    import ddtrace.auto  # noqa:F401
+
+    assert config._remote_config_enabled == rc_enabled
+
+    from ddtrace.internal.remoteconfig.worker import remoteconfig_poller
+
+    assert bool(remoteconfig_poller._client._products.get("APM_TRACING")) == rc_enabled
+    assert bool(remoteconfig_poller._client._products.get("AGENT_CONFIG")) == rc_enabled
+    assert bool(remoteconfig_poller._client._products.get("AGENT_TASK")) == rc_enabled
